api/requirements.txt
mysqlclient
Flask==2.2.2
scikit-learn
scipy==1.9.3
numpy
mysql-connector-python
